requirements.txt
# Bible Study AI Chat
flask>=3.0.0
python-dotenv>=1.0.0

# Gemini (Google) — used for Bible study chat
google-genai>=1.0.0

# Optional: other LLM backends
# openai>=1.0.0
# anthropic>=0.18.0
